macos porting, following the discussion on pull request
