+++ compat/linux/compat.h
@@ -1,0 +1,4 @@
+#ifndef COMPAT_H
+#define COMPAT_H
+
+#endif
