+++ compat/macos/endian.h
@@ -1,0 +1,4 @@
+#ifndef ENDIAN_H
+#define ENDIAN_H
+
+#endif
