+++ compat/macos/linux/if.h
@@ -1,0 +1,6 @@
+#ifndef IF_H
+#define IF_H
+
+#include <net/if.h>
+
+#endif
