+++ compat/macos/linux/if_packet.h
@@ -1,0 +1,12 @@
+#ifndef IF_PACKET_H
+#define IF_PACKET_H
+
+/* Ethernet */
+#define ETH_P_IP ETHERTYPE_IP
+#define ETH_P_ARP ETHERTYPE_ARP
+#define ETH_ALEN ETHER_ADDR_LEN
+
+/* for HW address */
+#define SIOCGIFHWADDR	0x8927		/* Get hardware address		*/
+
+#endif
