+++ compat/macos/netinet/ether.h
@@ -1,0 +1,9 @@
+#ifndef IF_NETINET_ETHER_H
+#define IF_NETINET_ETHER_H
+
+#include <net/ethernet.h>
+#include <arpa/inet.h>
+#include <net/if_dl.h>
+#include <net/bpf.h>
+
+#endif
