+++ vp/src/core/common/gdb-mc/register_format.cpp
@@ -1,12 +1,6 @@
 #include <iomanip>
 
-#ifdef __APPLE__
-#define bswap_16(value) ((((value) & 0xff) << 8) | ((value) >> 8))
-#define bswap_32(value) (((uint32_t)bswap_16((uint16_t)((value) & 0xffff)) << 16) | (uint32_t)bswap_16((uint16_t)((value) >> 16)))
-#define bswap_64(value) (((uint64_t)bswap_32((uint32_t)((value) & 0xffffffff)) << 32) | (uint64_t)bswap_32((uint32_t)((value) >> 32)))
-#else
 #include <byteswap.h>
-#endif
 
 #include "register_format.h"
