+++ vp/src/platform/common/abstract_uart.h
@@ -69,13 +69,9 @@ private:
 
 	std::queue<uint8_t> tx_fifo;
 	std::queue<uint8_t> rx_fifo;
-#ifdef __APPLE__
+
 	sem_t *rxempty_p;
 	sem_t *txfull_p;
-#else
-	sem_t rxempty;
-	sem_t txfull;
-#endif
 	vp::map::LocalRouter router = {"UART"};
 };
