+++ vp/src/platform/common/slip.cpp
@@ -11,36 +11,8 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 
-#ifdef __APPLE__
-#include <net/if.h>
-
-/* Ioctl defines */
-#define TUNSETNOCSUM  _IOW('T', 200, int) 
-#define TUNSETDEBUG   _IOW('T', 201, int) 
-#define TUNSETIFF     _IOW('T', 202, int) 
-#define TUNSETPERSIST _IOW('T', 203, int) 
-#define TUNSETOWNER   _IOW('T', 204, int)
-#define TUNSETLINK    _IOW('T', 205, int)
-#define TUNSETGROUP   _IOW('T', 206, int)
-#define TUNGETFEATURES _IOR('T', 207, unsigned int)
-#define TUNSETOFFLOAD  _IOW('T', 208, unsigned int)
-#define TUNSETTXFILTER _IOW('T', 209, unsigned int)
-#define TUNGETIFF      _IOR('T', 210, unsigned int)
-#define TUNGETSNDBUF   _IOR('T', 211, int)
-#define TUNSETSNDBUF   _IOW('T', 212, int)
-
-/* TUNSETIFF ifr flags */
-#define IFF_TUN                0x0001
-#define IFF_TAP                0x0002
-#define IFF_NO_PI        0x1000
-#define IFF_ONE_QUEUE        0x2000
-#define IFF_VNET_HDR        0x4000
-#define IFF_TUN_EXCL        0x8000
-
-#else
 #include <linux/if.h>
 #include <linux/if_tun.h>
-#endif 
 
 #include "slip.h"
